Database task fixes
Fixed apparent problems.
Rearranged code.
