+++ .gitignore
@@ -2,6 +2,9 @@
 #Ignore build files
 bin
 
+#Ignore DB info
+*.env
+
 ### Go ###
 # Binaries for programs and plugins
 *.exe
