--- pkg/app/dbdata.go
@@ -1,10 +1,0 @@
-package app
-
-const (
-	host   = "localhost"
-	port   = "5432"
-	user   = "rtgc"
-	pass   = "rtgcpass"
-	dbname = "rtgc"
-	ssl    = "disable"
-)
