+++ pkg/db/internal/rtgc/public/criteria.go
@@ -1,0 +1,13 @@
+package db
+
+type RoomsCriteria struct {
+	ID   *int
+	Name *string
+}
+
+type UsersInRoomCriteria struct {
+	ID       *int
+	RoomID   *int
+	UserID   *string
+	Username *string
+}
