+++ pkg/db/internal/rtgc/public/db.go
@@ -1,0 +1,22 @@
+package db
+
+import (
+	"context"
+	"database/sql"
+	"fmt"
+)
+
+func NewDB(dataSourceName string) (*sql.DB, error) {
+	db, err := sql.Open("postgres", dataSourceName)
+	if err != nil {
+		return nil, fmt.Errorf("newDB: %w", err)
+	}
+
+	return db, nil
+}
+
+type SQLHandler interface {
+	ExecContext(context.Context, string, ...interface{}) (sql.Result, error)
+	QueryContext(context.Context, string, ...interface{}) (*sql.Rows, error)
+	QueryRowContext(context.Context, string, ...interface{}) *sql.Row
+}
