+++ pkg/db/internal/rtgc/public/error.go
@@ -1,4 +1,4 @@
-package repository
+package db
 
 import (
 	"errors"
