+++ pkg/db/internal/rtgc/public/list.go
@@ -1,4 +1,4 @@
-package repository
+package db
 
 type SortDirection string
