+++ pkg/db/roomrepository/repository/rooms.go
@@ -3,13 +3,9 @@ package repository
 import (
 	"context"
 
+	db "github.com/nndergunov/RTGC-Project/pkg/db/internal/rtgc/public"
 	"github.com/nndergunov/RTGC-Project/pkg/db/internal/rtgc/public/model"
 )
-
-type RoomsCriteria struct {
-	ID   *int
-	Name *string
-}
 
 type RoomsRepository interface {
 	CreateRooms(ctx context.Context, name string) (id int, err error)
@@ -17,5 +13,5 @@ type RoomsRepository interface {
 	UpdateRooms(ctx context.Context, room *model.Rooms) error
 	DeleteRooms(ctx context.Context, id int) error
 
-	ListRooms(ctx context.Context, list *ListOptions, criteria *RoomsCriteria) ([]*model.Rooms, error)
+	ListRooms(ctx context.Context, list *db.ListOptions, criteria *db.RoomsCriteria) ([]*model.Rooms, error)
 }
