+++ pkg/domain/room.go
@@ -1,15 +1,9 @@
 package domain
 
 import (
-	"errors"
 	"fmt"
 	"log"
 	"time"
-)
-
-var (
-	errID    = errors.New("user with such ID already is in the room using valid username")
-	errUname = errors.New("such username is already taken in this room")
 )
 
 type Room struct {
