--- pkg/domain/user.go
@@ -1,6 +1,0 @@
-package domain
-
-type User struct {
-	ID    string
-	Chats []string
-}
